refactor output iter
