+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/worker/DPThread.scala
@@ -140,7 +140,7 @@ class DPThread(
       var channelId: ChannelIdentity = null
       var msgOpt: Option[WorkflowFIFOMessage] = None
       if (
-        dp.inputManager.hasUnfinishedInput || dp.outputManager.hasUnfinishedOutput || dp.pauseManager.isPaused
+        dp.inputManager.hasUnfinishedInput || dp.tupleProcessingManager.hasUnfinishedOutput || dp.pauseManager.isPaused
       ) {
         dp.inputGateway.tryPickControlChannel match {
           case Some(channel) =>
